backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio==5.12.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
bcrypt==4.2.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
httpx==0.28.1
